+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Changed
 
+- A mix of empty and non-empty `skills` arrays is allowed in input (#460)
 - Formatting script updated to use version 10 of clang-format (#452)
 - vroom will now read json input from stdin if no other input is specified (#457)
 - Clearer error message with invalid json response from http routing request (#471)